Implementation from back-porting code
Back-ported code changes to the "generic-item-page-field.component.ts"
file from the dspace/dspace-angular "main" branch".

Changed "Abstract" to use a "ds-generic-item-page-field" to control
Markdown rendering (turning if off), and enabling Markdown rendering
for the "dc.description" field.

These changes, as they are in the dspace-angular "main" branch, are
expected to be in DSpace 7.6