LIBDRUM-838. Checkpoint for 80/20 solution
Here's what I've got so far regarding a "Submit item to DRUM" link
(LIBDRUM-839).

For logged-in users, it is fairly straightforward to add a
"Submit item to DRUM" link in the navigation bar that acts the same way
as the "New | Item" entry in the administrative sidebar, i.e., displays
the "New item" modal dialog that prompts the user to select a collection
and then redirects them to the submission form.

The "Submit item to DRUM" link can be set to display in the navigation
bar when the user is logged in, and has "submit" permissions to at
least one collection. So logged-in users that can submit would see it,
but anonymous users (or users who are logged in, but can't submit) would
not see it.

I view this as an "80/20" solution -- it gets you most of what I think
we're looking for without a lot of effort (and keeping pretty much
within the bounds of what DSpace is expecting us to do).

This solution doesn't do anything for anonymous users. It looks like
it might be possible to still show the "Submit item to DRUM" link for
those anonymous users, and present a different prompt such as
"Please login to DRUM first", or something to that effect. However,
then we would have to distinguish between anonymous users (prompting
them to login) vs. logged-in users that can't submit (which would
need a different message). So displaying "Submit item to DRUM" for
anonymous users starts to get complicated. If you are interested in
this, I can investigate further.

https://umd-dit.atlassian.net/browse/LIBDRUM-839