LIBDRUM-874. Modified server config to support read-only container
Modified the "server.ts" file to place the AppConfig in an
"assets/config.json" directory.

This was done so that on startup the stock
"/app/dist/browser/assets/config.json" file would be written to
"/app/dist/writeable/assets/config.json" instead.

This makes it possible to have the "/apps/dist/browser/assets" directory,
which has other files in it be read-only, and the startup can still
write the file as expectd.

https://umd-dit.atlassian.net/browse/LIBDRUM-874